requirements.txt
flask
tensorflow
scikit-learn
pillow
numpy
joblib
